Issue #46: Rename and fix the formula classes for ock_preset.
